// src/shared/utils/types/enums.ts
export enum ERROR_CODES {
  // COMMON
  EMAIL_SERVICE_ERROR = "EMAIL_SERVICE_ERROR",
  INTERNAL_ERROR = "INTERNAL_ERROR",

  VALIDATION_ERROR = "VALIDATION_ERROR",
  INVALID_REQUEST = "INVALID_REQUEST",
  FORBIDDEN = "FORBIDDEN",
  CREDENTIAL_NOT_FOUND = "CREDENTIAL_NOT_FOUND",
  INVALID_TOKEN = "INVALID_TOKEN",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  ACCOUNT_BLOCKED = "ACCOUNT_BLOCKED",

  KAFKA_CONNECTION_FAILED = "KAFKA_CONNECTION_FAILED",
  KAFKA_SUBSCRIBE_FAILED = "KAFKA_SUBSCRIBE_FAILED",
  KAFKA_DISCONNECT_FAILED = "KAFKA_DISCONNECT_FAILED",
  KAFKA_CONSUMER_FAILED = "KAFKA_CONSUMER_FAILED",
  KAFKA_PRODUCER_CONNECT_FAILED = "KAFKA_PRODUCER_CONNECT_FAILED",
  KAFKA_PUBLISH_FAILED = "KAFKA_PUBLISH_FAILED",
  KAFKA_PRODUCER_DISCONNECT_FAILED = "KAFKA_PRODUCER_DISCONNECT_FAILED",

  DB_CONNECTION_FAILED = "DB_CONNECTION_FAILED",
  DB_DISCONNECT_FAILED = "DB_DISCONNECT_FAILED",

  PAYMENT_INVALID_RESPONSE = "PAYMENT_INVALID_RESPONSE",
  PAYMENT_BAD_REQUEST = "PAYMENT_BAD_REQUEST",
  PAYMENT_SERVICE_ERROR = "PAYMENT_SERVICE_ERROR",
  PAYMENT_SERVICE_UNAVAILABLE = "PAYMENT_SERVICE_UNAVAILABLE",

  // AUTH
  UNAUTHORIZED = "UNAUTHORIZED",
  TOKEN_EXPIRED = "TOKEN_EXPIRED",
  TOKEN_INVALID = "TOKEN_INVALID",

  // USER
  USER_NOT_FOUND = "USER_NOT_FOUND",
  PROVIDER_ALREADY_VERIFIED = "PROVIDER_ALREADY_VERIFIED",
  PROVIDER_PROFILE_UPDATE_FAILED = "PROVIDER_PROFILE_UPDATE_FAILED",

  // BOOKING
  BOOKING_NOT_FOUND = "BOOKING_NOT_FOUND",
  BOOKING_NOT_CONFIRMED = "BOOKING_NOT_CONFIRMED",

  // GOOGLE
  GOOGLE_TOKEN_MISSING = "GOOGLE_TOKEN_MISSING",
  GOOGLE_API_ERROR = "GOOGLE_API_ERROR",
  TOKEN_MISSING = "TOKEN_MISSING",

  // ENCRYPTION
  AES_DECRYPT_FAIL = "AES_DECRYPT_FAIL",
  AES_ENCRYPT_FAIL = "AES_ENCRYPT_FAIL",

  // RESOURCE NOT FOUND
  REVIEW_NOT_FOUND = "REVIEW_NOT_FOUND",
  SUBSCRIPTION_NOT_FOUND = "SUBSCRIPTION_NOT_FOUND",
  PROVIDER_PROFILE_NOT_FOUND = "PROVIDER_PROFILE_NOT_FOUND",
  PROVIDER_SERVICE_NOT_FOUND = "PROVIDER_SERVICE_NOT_FOUND",
  SERVICE_AVAILABILITY_NOT_FOUND = "SERVICE_AVAILABILITY_NOT_FOUND",

  // SLOT
  SLOT_NOT_FOUND = "SLOT_NOT_FOUND",
  SLOT_NOT_AVAILABLE = "SLOT_NOT_AVAILABLE",

  // ADDRESS
  ADDRESS_NOT_FOUND = "ADDRESS_NOT_FOUND",

  // USER
  FAILED_TO_UPDATE_USER = "FAILED_TO_UPDATE_USER",

  // SUBSCRIPTION
  SUBSCRIPTION_ALREADY_LIVE = "SUBSCRIPTION_ALREADY_LIVE",

  // SERVICE
  SERVICE_NOT_FOUND = "SERVICE_NOT_FOUND",
  SERVICE_ALREADY_EXIST = "SERVICE_ALREADY_EXISTS",

  // PLAN
  PLAN_NOT_FOUND = "PLAN_NOT_FOUND",
  PLAN_ALREADY_EXIST = "PLAN_ALREADY_EXIST",
}

export enum IdType {
  EVENT = "EVENT",
  TRANSACTION = "TRANSACTION",
  ROOM = "ROOM",
  IDEMPOTENCY = "IDEMPOTENCY",
  FILE = "FILE",
}